requirements.txt
textual
redis
rich
python-dotenv
elasticsearch<8
